requirements.txt
pandas==1.0.3
praat-parselmouth==0.3.3
pympi-ling==1.69
scikit-learn==0.22.1
glob2==0.7
numpy==1.18.1